api/requirements.txt
gunicorn==22.0.0
Flask==3.0.3
Flask-Cors==4.0.0
kubernetes==29.0.0
numpy==1.26.0
oauthlib==3.2.2
pytz==2024.1
randomname==0.2.1
rsa==4.9
scipy==1.11.3
biopython==1.84
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
requests-mock==1.12.1
pydantic==2.10.6
PyJWT==2.10.1